requirements.txt
absl-py==1.4.0
dnspython==2.3.0
numpy==1.24.3
ortools==9.6.2534
protobuf==4.23.1
psutil==5.9.5
pymongo==4.3.3
python-dotenv==1.0.0
pytz==2023.3
scipy==1.10.1
pandas==2.2.0
requests==2.31.0
